% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_cogs.R
\name{auto_cogs}
\alias{auto_cogs}
\title{Compute automatic cognostics}
\usage{
auto_cogs(data)
}
\arguments{
\item{data}{a list of data frames (one per subset), a grouped data frame, or a nested data frame}
}
\value{
If the input is a list of data frames, the return value is a list of data frames containing the cognostics. If the input is a grouped or nested df, the result will be a nested df with a new column containing the cognostics.
}
\description{
Compute automatic cognostics
}
\seealso{
\code{\link{trelliscope}}
}
