% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{by_row_cog}
\alias{by_row_cog}
\title{Apply a function to each row of a data frame and return a data frame with new column of cognostics}
\usage{
by_row_cog(.d, ..f, .to = NULL)
}
\arguments{
\item{.d}{A data frame.}

\item{..f}{A function to apply to each row. It should return a single-row data frame of cognostics}

\item{.to}{Name of output column (defaults to "cogs" or if "cogs" exists, "cogs1", etc.).}
}
\description{
Apply a function to each row of a data frame and return a data frame with new column of cognostics
}
\details{
See \code{\link[purrrlyr]{by_row}}
}
\examples{
\dontrun{
library(tidyr)
library(purrr)
library(rbokeh)
iris \%>\%
  nest(-Species) \%>\%
  mutate(mod = map(data, ~ lm(Sepal.Length ~ Sepal.Width, data = .x))) \%>\%
  panels_by_row(function(x) {
    figure(xlab = "Sepal.Width", ylab = "Sepal.Length") \%>\%
      ly_points(x$data[[1]]$Sepal.Width, x$data[[1]]$Sepal.Length) \%>\%
      ly_abline(x$mod[[1]])
  }) \%>\%
  trelliscope(name = "iris")
}
}
