% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cogs}
\alias{cogs}
\title{Cogs Wrapper Function
Wrapper function to specify a cognostics function to be applied to a list-column for use tidy group/nest/mutate/map-like situations}
\usage{
cogs(.x, .f, ...)
}
\arguments{
\item{.x}{a list or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{.f}{a function, formula, or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{...}{additional arguments passed on to .f (see \code{\link[purrr]{map}} for details)}
}
\description{
Cogs Wrapper Function
Wrapper function to specify a cognostics function to be applied to a list-column for use tidy group/nest/mutate/map-like situations
}

