% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{copy_viewer_files}
\alias{copy_viewer_files}
\title{Copy Trelliscope viewer files to local app directory}
\usage{
copy_viewer_files(base_path,
  src = "https://raw.githubusercontent.com/hafen/trelliscopejs-demo/gh-pages/")
}
\arguments{
\item{base_path}{the base directory of the trelliscope application}

\item{src}{the location of the Trelliscope viewer files}
}
\description{
Copy Trelliscope viewer files to local app directory
}

