% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{create_cog_template}
\alias{create_cog_template}
\title{Create a cognostics template that can be edited and used to specify
cognostics in a display}
\usage{
create_cog_template(x = NULL)
}
\arguments{
\item{x}{a data frame that will be used as an input for the
trelliscope display. If NULL, a blank template will be created.}
}
\description{
Create a cognostics template that can be edited and used to specify
cognostics in a display
}
\note{
The input `x` can be a starting point and does not need to
contain all variables that may be added later.
Also note that after you edit the output (for example using the
editData package, or writing to a csv and editing in Excel) to reflect
how you would like the cognostics to display in the viewer, you can
Add this specification to your trelliscope display with
[add_cog_template()].
}
