% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_promise.R
\name{get_ppanel}
\alias{get_ppanel}
\title{Get an evaluated panel promise for a specified row index}
\usage{
get_ppanel(dat, col_name = "panel", row = 1)
}
\arguments{
\item{dat}{a data frame with a column containing a panel promise}

\item{col_name}{the name of the column containing the panel promise}

\item{row}{the row index to use to find the panel to evaluate}
}
\description{
Get an evaluated panel promise for a specified row index
}
