% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{img_panel}
\alias{img_panel}
\title{Cast a vector of URLs pointing to images as an image panel source}
\usage{
img_panel(x)
}
\arguments{
\item{x}{a vector of URLs pointing to images}
}
\description{
Cast a vector of URLs pointing to images as an image panel source
}

