% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cogs.R
\name{input_cogs}
\alias{input_cogs}
\title{Specify a collection of input cognostics}
\usage{
input_cogs(...)
}
\arguments{
\item{\ldots}{objects created by any of \code{\link{input_radio}},
\code{\link{input_text}}, \code{\link{input_textarea}}}
}
\description{
Specify a collection of input cognostics
}
