% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cogs.R
\name{input_radio}
\alias{input_radio}
\title{Specify a radio button input}
\usage{
input_radio(name, desc = NULL, options, group = NULL, default_label = TRUE)
}
\arguments{
\item{name}{name of the input}

\item{desc}{optional description of the input}

\item{options}{a character vector of options to select between}

\item{group}{optional categorization of the input for organizational purposes in the viewer (currently not implemented in the viewer)}

\item{default_label}{should this input be shown under the panel in the viewer by default?}
}
\description{
Specify a radio button input
}
