% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_cogs.R
\name{input_text}
\alias{input_text}
\title{Specify a text input}
\usage{
input_text(
  name,
  desc = NULL,
  width = 80,
  height = 3,
  group = NULL,
  default_label = TRUE
)
}
\arguments{
\item{name}{name of the input}

\item{desc}{optional description of the input}

\item{width}{width (in characters) of the text box popout}

\item{height}{height (in lines of text) of the text box popout}

\item{group}{optional categorization of the input for organizational purposes in the viewer (currently not implemented in the viewer)}

\item{default_label}{should this input be shown under the panel in the viewer by default?}
}
\description{
Specify a text input
}
