% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_promise.R
\name{map2_pplot}
\alias{map2_pplot}
\title{Map over multiple inputs simultaneously and return a vector of plots}
\usage{
map2_pplot(.x, .y, .f, ...)
}
\arguments{
\item{.x, .y}{Vectors of the same length. A vector of length 1 will be recycled.}

\item{.f}{A function, formula, or atomic vector (see \code{\link[purrr]{map2}} for details)}

\item{...}{additional arguments passed on to .f.}

\item{.l}{A list of lists. The length of .l determines the number of arguments that .f will be called with. List names will be used if present.}
}
\description{
Map over multiple inputs simultaneously and return a vector of plots
}
\details{
See \code{\link[purrr]{map2}}
}
\examples{
\dontrun{
library(tidyr)
library(purrr)
library(rbokeh)
library(dplyr)

p <- iris \%>\%
  nest(-Species) \%>\%
  mutate(
    mod = map(data, ~ lm(Sepal.Length ~ Sepal.Width, data = .x)),
    panel = map2_pplot(data, mod, function(data, mod) {
      figure(xlab = "Sepal.Width", ylab = "Sepal.Length") \%>\%
        ly_points(data$Sepal.Width, data$Sepal.Length) \%>\%
        ly_abline(mod)
    }))

# look at one panel
p$panel[[1]]
get_ppanel(p, row = 2)

# plot it
trelliscope(p, name = "iris")

# same plot but using pmap_pplot()
p <- iris \%>\%
  nest(-Species) \%>\%
  mutate(
    mod = map(data, ~ lm(Sepal.Length ~ Sepal.Width, data = .x)),
    panel = pmap_pplot(list(data = data, mod = mod), function(data, mod) {
      figure(xlab = "Sepal.Width", ylab = "Sepal.Length") \%>\%
        ly_points(data$Sepal.Width, data$Sepal.Length) \%>\%
        ly_abline(mod)
    }))

# look at one panel
p$panel[[1]]
get_ppanel(p, row = 2)

# plot it
trelliscope(p, name = "iris")
}
}
