% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{md_description}
\alias{md_description}
\title{Specify a markdown description for a display}
\usage{
md_description(
  content = "",
  title = "Information About This Display",
  show = FALSE
)
}
\arguments{
\item{content}{Markdown content}

\item{title}{Title of the dialog box that displays this content}

\item{show}{should the markdown description be shown by default when the display is loaded?}
}
\description{
Specify a markdown description for a display
}
