% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs-package.R
\docType{data}
\name{mpg_labels}
\alias{mpg_labels}
\title{Labels for ggplot2 "mpg" data}
\format{An object of class \code{list} of length 10.}
\usage{
mpg_labels
}
\description{
Labels for ggplot2 "mpg" data
}
\keyword{datasets}

