% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{panel}
\alias{panel}
\title{Panel Wrapper Function
Wrapper function to specify a plot object for a panel for use in dplyr summarise()}
\usage{
panel(x)
}
\arguments{
\item{x}{a plot object}
}
\description{
Panel Wrapper Function
Wrapper function to specify a plot object for a panel for use in dplyr summarise()
}
\examples{
\dontrun{
library(rbokeh)
library(dplyr)
ggplot2::mpg \%>\%
  group_by(manufacturer, class) \%>\%
  summarise(
    panel = panel(
      figure(xlab = "City mpg", ylab = "Highway mpg") \%>\%
        ly_points(cty, hwy)))
}
}

