% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{panels}
\alias{panels}
\title{Panels Wrapper Function
Wrapper function to specify a plot command to be applied to a list-column for use tidy group/nest/mutate/map-like situations}
\usage{
panels(.x, .f, ...)
}
\arguments{
\item{.x}{a list or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{.f}{a function, formula, or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{...}{additional arguments passed on to .f (see \code{\link[purrr]{map}} for details)}
}
\description{
Panels Wrapper Function
Wrapper function to specify a plot command to be applied to a list-column for use tidy group/nest/mutate/map-like situations
}
\examples{
mpg \%>\%
  group_by(manufacturer, class) \%>\%
  nest() \%>\%
  mutate(panel = panels(data,
    ~ figure(xlab = "City mpg", ylab = "Highway mpg") \%>\%
        ly_points(cty, hwy, data = .x))) \%>\%
  trelliscope(name = "city_vs_highway_mpg")
}

