% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.R
\name{panels_by_row}
\alias{panels_by_row}
\title{Panels Wrapper Function
Wrapper function to specify a plot command to be applied to a list-column for use tidy group/nest/mutate/map-like situations}
\usage{
panels_by_row(.d, ..f, .to = "panel")
}
\arguments{
\item{.d}{A data frame.}

\item{..f}{A function to apply to each row. It should return a valid panel object (such as a ggplot2 / lattice / htmlwidget object).}

\item{.to}{Name of output column (defaults to "panel").}
}
\description{
Panels Wrapper Function
Wrapper function to specify a plot command to be applied to a list-column for use tidy group/nest/mutate/map-like situations
}
\examples{
\dontrun{
library(tidyr)
library(purrr)
library(rbokeh)
iris \%>\%
  nest(-Species) \%>\%
  mutate(mod = map(data, ~ lm(Sepal.Length ~ Sepal.Width, data = .x))) \%>\%
  panels_by_row(function(x) {
    figure(xlab = "Sepal.Width", ylab = "Sepal.Length") \%>\%
      ly_points(x$data[[1]]$Sepal.Width, x$data[[1]]$Sepal.Length) \%>\%
      ly_abline(x$mod[[1]])
  }) \%>\%
  trelliscope(name = "iris")
}
}

