% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{+.gg}
\alias{+.gg}
\title{Add method to transfer the}
\usage{
\method{+}{gg}(e1, e2)
}
\arguments{
\item{e1}{a object with class gg}

\item{e2}{if object is of class 'facet_trelliscope', then 'facet_trelliscope' will be appended to the class of e1}
}
\description{
Add method to transfer the
}

