% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{prepare_display}
\alias{prepare_display}
\title{Set up all auxilliary files needed for a Trelliscope app}
\usage{
prepare_display(base_path, id, jsonp = TRUE)
}
\arguments{
\item{base_path}{the base directory of the trelliscope application}

\item{id}{a unique id for the application}

\item{jsonp}{should json for display list and app config be jsonp (TRUE) or json (FALSE)?}
}
\description{
Set up all auxilliary files needed for a Trelliscope app
}

