% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{print.facet_trelliscope}
\alias{print.facet_trelliscope}
\title{Print facet trelliscope object}
\usage{
\method{print}{facet_trelliscope}(x, ...)
}
\arguments{
\item{x}{plot object}

\item{...}{ignored}
}
\description{
Print facet trelliscope object
}
