% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{sort_spec}
\alias{sort_spec}
\title{Specify how a display should be sorted}
\usage{
sort_spec(name, dir = "asc")
}
\arguments{
\item{name}{variable name to sort on}

\item{dir}{direction to sort ('asc' or 'desc')}
}
\description{
Specify how a display should be sorted
}
