% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charm.R
\name{tr_charm}
\alias{tr_charm}
\title{Use fidelius to password protect a trelliscope display}
\usage{
tr_charm(x, ...)
}
\arguments{
\item{x}{an object of class "facet_trelliscope" or 
"trelliscopejs_widget"}

\item{...}{arguments passed to [fidelius::charm()]}
}
\description{
Use fidelius to password protect a trelliscope display
}
