% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope.R
\name{trelliscope}
\alias{trelliscope}
\title{Create a Trelliscope Display}
\usage{
trelliscope(x, name, group = "common", desc = "", cond_cols = NULL,
  md_desc = "", path, height = 500, width = 500, state = NULL,
  jsonp = TRUE)
}
\arguments{
\item{x}{an object to create at trelliscope display for}

\item{name}{name of the display}

\item{group}{group that the display belongs to}

\item{desc}{description of the display}

\item{cond_cols}{the column name(s) that comprise the conditioning variables}

\item{md_desc}{optional string of markdown that will be shown in the viewer for additional context about the display}

\item{path}{the base directory of the trelliscope application}

\item{height}{height in pixels of each panel}

\item{width}{width in pixels of each panel}

\item{state}{the initial state the display will open in}

\item{jsonp}{should json for display object be jsonp (TRUE) or json (FALSE)?}
}
\description{
Create a Trelliscope Display
}
\examples{
\dontrun{
library(dplyr)
library(rbokeh)

dd <- iris \%>\%
  group_by(Species) \%>\%
  summarise(
    msl = cog(mean(Sepal.Length), desc = "mean sepal length"),
    panel = panel(figure() \%>\%
      ly_points(Sepal.Length, Sepal.Width))) \%>\%
  trelliscope(name = "test", cond_cols = "Species")

dd
}
}

