% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope2-package.R
\docType{package}
\name{trelliscope2-package}
\alias{trelliscope2-package}
\title{trelliscope2}
\description{
Create interactive Trelliscope displays
}
\details{
\url{http://hafen.github.io/trelliscope2/}
}
\examples{
help(package = trelliscope2)
}

