% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope.R
\name{update_display_list}
\alias{update_display_list}
\title{Update Trelliscope app display list file}
\usage{
update_display_list(base_path, jsonp = TRUE)
}
\arguments{
\item{base_path}{the base directory of the trelliscope application}

\item{jsonp}{should json for display list be jsonp (TRUE) or json (FALSE)?}
}
\description{
Update Trelliscope app display list file
}

