% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{view_display}
\alias{view_display}
\title{View a Trelliscope Display}
\usage{
view_display(base_path)
}
\arguments{
\item{base_path}{the base directory of the trelliscope application}
}
\description{
View a Trelliscope Display
}

