% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{view_item}
\alias{view_item}
\title{Construct a list of view items}
\usage{
view_item(name, hash)
}
\arguments{
\item{name}{A string indicating the name of the view (will be displayed in the "Views" sidebar)}

\item{hash}{A URL hash that sends the user to this view (typically the URL starting with &nrow=...)}
}
\description{
Construct a list of view items
}
