% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{view_list}
\alias{view_list}
\title{Construct a list of view items}
\usage{
view_list(...)
}
\arguments{
\item{...}{Objects created using \code{\link{view_item}}}
}
\description{
Construct a list of view items
}
