% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{write_cognostics}
\alias{write_cognostics}
\title{Write cognostics data for a display in a Trelliscope app}
\usage{
write_cognostics(cogdf, base_path, id, name, group = "common",
  jsonp = TRUE)
}
\arguments{
\item{cogdf}{a data frame of cognostics, prepared with \code{\link{as_cognostics}}}

\item{base_path}{the base directory of the trelliscope application}

\item{id}{a unique id for the application}

\item{name}{name of the display}

\item{group}{group that the display belongs to}

\item{jsonp}{should json for cognostics be jsonp (TRUE) or json (FALSE)?}
}
\description{
Write cognostics data for a display in a Trelliscope app
}
