% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{write_display_obj}
\alias{write_display_obj}
\title{Write a "display object" file for a Trelliscope app}
\usage{
write_display_obj(cogdf, panel_example, base_path, id, name, group = "common",
  desc = "", height = 500, width = 500, md_desc = "", state = NULL,
  jsonp = TRUE)
}
\arguments{
\item{cogdf}{a data frame of cognostics, prepared with \code{\link{as_cognostics}}}

\item{panel_example}{an example object of one panel of a display (can be trellis, ggplot2, or htmlwidget object)}

\item{base_path}{the base directory of the trelliscope application}

\item{id}{a unique id for the application}

\item{name}{name of the display}

\item{group}{group that the display belongs to}

\item{desc}{description of the display}

\item{height}{height in pixels of each panel}

\item{width}{width in pixels of each panel}

\item{md_desc}{optional string of markdown that will be shown in the viewer for additional context about the display}

\item{state}{the initial state the display will open in}

\item{jsonp}{should json for display object be jsonp (TRUE) or json (FALSE)?}
}
\description{
Write a "display object" file for a Trelliscope app
}

