% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscope.R
\name{write_panel}
\alias{write_panel}
\title{Write a plot object as a panel in a Trelliscope display}
\usage{
write_panel(plot_object, key, base_path, name, group = "common", width,
  height, jsonp = TRUE)
}
\arguments{
\item{plot_object}{a plot object to be written (can be trellis, ggplot2, or htmlwidget)}

\item{key}{a string identifying the panel key, which will be used as the panel file name and which the \code{panelKey} column of the cognostics data frame should point to}

\item{base_path}{the base directory of the trelliscope application}

\item{name}{name of the display that the panel belongs to}

\item{group}{group name of the display that the panel belongs to}

\item{width}{width in pixels of each panel}

\item{height}{height in pixels of each panel}

\item{jsonp}{should json for panel be jsonp (TRUE) or json (FALSE)?}
}
\description{
Write a plot object as a panel in a Trelliscope display
}

