% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{write_panels}
\alias{write_panels}
\title{Write a list of plot objects as panels in a Trelliscope display}
\usage{
write_panels(plot_list, ..., pb = NULL)
}
\arguments{
\item{plot_list}{a named list of plot objects to be written as panels (objects can be trellis, ggplot2, or htmlwidget) with the list names being the keys for the panels}

\item{...}{params passed directly to \code{\link{write_panel}}}

\item{pb}{optional progress bar object to pass in and use to report progress}
}
\description{
Write a list of plot objects as panels in a Trelliscope display
}
