% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{as_cognostics}
\alias{as_cognostics}
\title{Cast a data frame as a cognostics data frame}
\usage{
as_cognostics(x, cond_cols, key_col = NULL)
}
\arguments{
\item{x}{a data frame}

\item{cond_cols}{the column name(s) that comprise the conditioning variables}

\item{key_col}{the column name that indicates the panel key}
}
\description{
Cast a data frame as a cognostics data frame
}

