% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{as_cognostics}
\alias{as_cognostics}
\title{Cast a data frame as a cognostics data frame}
\usage{
as_cognostics(x, cond_cols, key_col = NULL, cog_desc = NULL,
  needs_key = TRUE, needs_cond = TRUE, group = "common")
}
\arguments{
\item{x}{a data frame}

\item{cond_cols}{the column name(s) that comprise the conditioning variables}

\item{key_col}{the column name that indicates the panel key}

\item{cog_desc}{an optional named list of descriptions for the cognostics columns}

\item{needs_key}{does the result need to have a "key" column?}

\item{needs_cond}{does the result need to have conditioning variable columns?}

\item{group}{value to be used in the \code{\link{cog}} group}
}
\description{
Cast a data frame as a cognostics data frame
}
