% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cog.R
\name{cog_href}
\alias{cog_href}
\title{Href Cognostic}
\usage{
cog_href(
  x,
  desc = "link",
  group = "common",
  type = c("href", "href_hash"),
  default_label = FALSE,
  default_active = FALSE,
  filterable = FALSE,
  sortable = FALSE,
  log = FALSE
)
}
\arguments{
\item{x}{URL to link to}

\item{desc, group, default_label, default_active, filterable, sortable, log}{arguments passed to \code{\link{cog}}}

\item{type}{of either "href" or "href_hash". "href" will open the link in a new page. "href_hash" will update the page's hash and reload the page (useful when changing state inside an iframe)}
}
\description{
Create href to be used as cognostics in a trelliscope display
}
\examples{
\donttest{
library(dplyr)
library(rbokeh)
iris \%>\%
  group_by(Species) \%>\%
  summarise(
    wiki_link = cog_href(paste0("https://en.wikipedia.org/wiki/Iris_",
      tolower(Species))[1], default_label = TRUE,
      desc = "link to species on wikipedia"),
    panel = panel(figure(xlab = "Sepal Length", ylab = "Sepal Width") \%>\%
      ly_points(Sepal.Length, Sepal.Width))) \%>\%
  trelliscope(name = "iris_species", ncol = 3)
}
}
\seealso{
\code{\link{cog}}
}
