% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{cogs}
\alias{cogs}
\title{Cogs Wrapper Function
Wrapper function to specify a cognostics function to be applied to a list-column for use tidy group/nest/mutate/map-like situations}
\usage{
cogs(.x, .f, ...)
}
\arguments{
\item{.x}{a list or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{.f}{a function, formula, or atomic vector (see \code{\link[purrr]{map}} for details)}

\item{...}{additional arguments passed on to .f (see \code{\link[purrr]{map}} for details)}
}
\description{
Cogs Wrapper Function
Wrapper function to specify a cognostics function to be applied to a list-column for use tidy group/nest/mutate/map-like situations
}
\examples{
\dontrun{
library(dplyr)
library(tidyr)
ggplot2::mpg \%>\%
  group_by(manufacturer, class) \%>\%
  nest() \%>\%
  mutate(
    additional_cogs = cogs(data,
      ~ data_frame(
        max_city_mpg = cog(max(.x$cty), desc = "Max city mpg"),
        min_city_mpg = cog(min(.x$cty), desc = "Min city mpg"))),
    panel = panels(data, ~ figure(xlab = "City mpg", ylab = "Highway mpg") \%>\%
      ly_points(cty, hwy, data = .x))) \%>\%
  trelliscope(name = "city_vs_highway_mpg", nrow = 1, ncol = 2)
}
}

