% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{facet_trelliscope}
\alias{facet_trelliscope}
\title{Facet Trelliscope}
\usage{
facet_trelliscope(facets, nrow = 1, ncol = 1, scales = "same",
  name = NULL, group = "common", desc = ggplot2::waiver(),
  md_desc = ggplot2::waiver(), path = NULL, height = 500, width = 500,
  prerender = TRUE, state = NULL, jsonp = TRUE, as_plotly = FALSE,
  plotly_args = NULL, self_contained = FALSE, thumb = TRUE,
  auto_cog_data = TRUE, split_layout = FALSE)
}
\arguments{
\item{facets}{Formula to facet the panels on. Similar to \code{ggplot2::\link[ggplot2]{facet_wrap}}'s \code{facets}.}

\item{nrow}{The number of rows of panels to display by default.}

\item{ncol}{The number of columns of panels to display by default.}

\item{scales}{Should scales be the same (\code{"same"}, the default), free (\code{"free"}), or sliced (\code{"sliced"}). May provide a single string or two strings, one for the X and Y axis respectively.}

\item{name}{name of the display}

\item{group}{Group that the display belongs to.}

\item{desc}{Description of the display.}

\item{md_desc}{Optional string of markdown that will be shown in the viewer for additional context about the display.}

\item{path}{The base directory of the trelliscope application.}

\item{height}{Height in pixels of each panel.}

\item{width}{Width in pixels of each panel.}

\item{prerender}{Should the panels be prerendered? (see note)}

\item{state}{The initial state the display will open in.}

\item{jsonp}{Should json for display object be jsonp (TRUE) or json (FALSE)?}

\item{as_plotly}{Should the panels be written as plotly objects?}

\item{plotly_args}{Optional named list of arguments to send to \code{ggplotly}.}

\item{self_contained}{Should the Trelliscope display be a self-contained html document? (see note)}

\item{thumb}{Should a thumbnail be created?}

\item{auto_cog_data}{Boolean that determines if automatic cognostics are produces}

\item{split_layout}{boolean that determines if the layout is split into components like a facet_grid vs. individual panels like facet_wrap.  Only applies to ggplot2 plot objects.}
}
\description{
Facet Trelliscope
}
\note{
There is a tradeoff with prerendering panels. Prerendering has the benefit that the output is fully self-contained and can be easily shared with others, whereas when you don't prerender an R server must be running in the background to generate the panels on demand. However, especially with ggplot2, when there are many panels it can take a long time to prerender.

Note that \code{self_contained} is severely limiting and should only be used in cases where you would either like your display to show up in the RStudio viewer pane, in an interactive R Markdown Notebook, or in a self-contained R Markdown html document.
}
\examples{
\dontrun{
library(ggplot2)

# basically swap out facet_wrap for facet_trelliscope
qplot(cty, hwy, data = mpg) +
  facet_trelliscope(~ class + manufacturer)

# not required, but if you set labels, these will be added as
# descriptions to the cognostics that are automatically computed
mpg <- set_labels(mpg, mpg_labels)

qplot(cty, hwy, data = mpg) +
  theme_bw() +
  facet_trelliscope(~ manufacturer + class, nrow = 2, ncol = 4)

# using plotly
library(plotly)
qplot(cty, hwy, data = mpg) +
  theme_bw() +
  facet_trelliscope(~ manufacturer + class, nrow = 2, ncol = 4, as_plotly = TRUE)

qplot(class, cty, data = mpg, geom = c("boxplot", "jitter")) +
  facet_trelliscope(~ class, ncol = 7, height = 800, width = 200,
    state = list(sort = list(sort_spec("cty_mean")))) +
  theme_bw()
}
}
