% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{facet_trelliscope}
\alias{facet_trelliscope}
\title{Facet Trelliscope}
\usage{
facet_trelliscope(..., path = tempdir())
}
\arguments{
\item{...}{all parameters passed onto \code{ggplot2::\link[ggplot2]{facet_wrap}}}

\item{path}{the base directory of the trelliscope application}
}
\description{
Facet Trelliscope
}
\examples{
\dontrun{
library(ggplot2)
p <- qplot(cty, hwy, data = mpg) +
  facet_trelliscope(
    ~ class + manufacturer,
    path = "_test"
  )
p
}
}

