% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_trelliscope.R
\name{facet_trelliscope}
\alias{facet_trelliscope}
\title{Facet Trelliscope}
\usage{
facet_trelliscope(..., dir = tempdir())
}
\arguments{
\item{...}{all parameters passed onto \code{ggplot2::\link[ggplot]{facet_wrap}}}
}
\description{
Facet Trelliscope
}
\examples{
p <- qplot(cty, hwy, data = mpg) +
  facet_trelliscope(
    ~ class + manufacturer,
    dir = "_test"
  )
p
}

