% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_writers.R
\name{prepare_display}
\alias{prepare_display}
\title{Set up all auxilliary files needed for a Trelliscope app}
\usage{
prepare_display(base_path, jsonp = TRUE, copy_viewer_files = TRUE)
}
\arguments{
\item{base_path}{the base directory of the trelliscope application}

\item{jsonp}{should json for display list and app config be jsonp (TRUE) or json (FALSE)?}

\item{copy_viewer_files}{should the viewer files be copied (using \code{\link{copy_viewer_files}})}
}
\description{
Set up all auxilliary files needed for a Trelliscope app
}

