% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trelliscopejs-package.R
\docType{package}
\name{trelliscopejs-package}
\alias{trelliscopejs-package}
\title{trelliscopejs}
\description{
Create interactive Trelliscope displays
}
\details{
\url{http://hafen.github.io/trelliscopejs/}
}
\examples{
help(package = trelliscopejs)
}

